Moves activity cards to events page
